requirements.txt
# pyodbc==4.0.32
Flask==2.1.1
Flask-MySQLdb==1.0.1
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
click==8.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
mysqlclient==2.1.0
setuptools==62.1.0
zipp==3.8.0
python-dotenv~=0.20.0
gunicorn
pymysql
